create first model migration

+++ ihc/core/consumers.py
@@ -4,6 +4,9 @@ from channels.generic.websocket import WebsocketConsumer
 from ihc.core.models import Client, Group
 from asgiref.sync import async_to_sync
 
+ACTION = 0
+VALUE = 0
+
 
 class UserConsumer(WebsocketConsumer):
     client = None
@@ -15,12 +18,13 @@ class UserConsumer(WebsocketConsumer):
         self.accept()
 
     def disconnect(self, close_code):
-        if self.group.owner == self.client:
+        if self.group and self.group.owner == self.client:
             self.group.delete()
         self.client.delete()
 
     def receive(self, text_data=None, bytes_data=None):
-        if text_data == 'create_group':
+        data = text_data.split(':')
+        if data[ACTION] == 'create_group':
             group_name = Group.generate_valid_group_name()
             async_to_sync(self.channel_layer.group_add)(group_name, self.channel_name)
             self.group = Group.objects.create(name=group_name,
@@ -28,6 +32,13 @@ class UserConsumer(WebsocketConsumer):
             self.client.group = self.group
             self.client.save()
             self.send(text_data=json.dumps({'group_id': group_name}))
+        elif data[ACTION] == 'join_group':
+            self.group = Group.objects.filter(name=data[VALUE]).first()
+            self.client.group = self.group
+            self.client.save()
+        elif data[ACTION] == 'exit_group':
+            self.client.group = None
+            self.client.save()
 
     def send_message(self, message):
         # async_to_sync(self.channel_layer.send)(self.client.channel_ws, {
